refactoring: выделяет компонент табы, из карточки фильма
